HOCS-5148: Refactor message validation.
